% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_associations_data.R
\name{feature_associations_data}
\alias{feature_associations_data}
\title{Prepare topic signatures for training feature associations.}
\usage{
feature_associations_data(object, mods, normalize = TRUE)
}
\arguments{
\item{object}{A Seurat object containing modalities with signature matrices
in object[[modality]]@meta.data.}

\item{mods}{A vector of modality names to extract from the Seurat object.}

\item{normalize}{If TRUE, rows of the signature matrices are normalized to
sum to 1. (default: TRUE)}
}
\value{
A nested list where the keys are the names of the modalities. Each
 modality contains a two-element list. The first element is a
 torch.FloatTensor of shape (n_topics, n_features). The second element is
 a data frame corresponding to the signatures.
}
\description{
Extracts and optionally row-normalizes signature matrices from
 multiple modalities in a Seurat object. Each signature matrix is expected
 to be stored in meta data of the respective modality. Matrices are returned
 in a dictionary keyed by modality name.
}
