% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_var_list.R
\name{filter_var_list}
\alias{filter_var_list}
\title{Retain a specific list of features in a Seurat object.}
\usage{
filter_var_list(object, var)
}
\arguments{
\item{object}{A raw, unfiltered Seurat object to be processed.}

\item{var}{A vector of feature names to be retained in the Seurat object.}
}
\value{
A Seurat object containing only the specified features.
}
\description{
This function retains only the specified list of features
 (e.g., genes, proteins) in a Seurat object and removes all other features.
 It is designed to streamline downstream analysis by focusing on a
 predefined subset of relevant features.
}
\note{
Ensure the feature names in var match the names in the dataset to
 avoid errors.
}
