% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr_heatmap.R
\name{plot_corr_heatmap}
\alias{plot_corr_heatmap}
\title{Visualize the correlation matrix between two sets of features as
 a heatmap.}
\usage{
plot_corr_heatmap(
  data1,
  data2,
  label1 = NULL,
  label2 = NULL,
  cmap = "bwr",
  fontsize = 8,
  figsize = c(8, 6),
  transparent = FALSE,
  save = NULL
)
}
\arguments{
\item{data1}{A data frame representing the first set of features. Each column
corresponds to a feature.}

\item{data2}{A data frame representing the second set of features. Each
column corresponds to a feature.}

\item{label1}{Label for the y-axis, representing the features from arr1.
If NULL, no label is set. Default is NULL.}

\item{label2}{Label for the x-axis, representing the features from arr2.
If NULL, no label is set. Default is NULL.}

\item{cmap}{Colormap for the heatmap. Default is ‘bwr’ (blue-white-red
colormap)}

\item{fontsize}{Font size for axis labels and colorbar ticks. Default is 8.}

\item{figsize}{Vector specifying the figure size c(width, height) in inches.
Default is c(8, 6).}

\item{transparent}{If TRUE, saves the figure with a transparent background.
Default is FALSE.}

\item{save}{File path to save the figure. If NULL, the figure is displayed
but not saved. Default is NULL.}
}
\description{
This function computes and plots a correlation heatmap to
 visualize the relationships between two sets of features. Each set is
 represented as data frame, with columns as features. The color intensity
 in the heatmap indicates the strength and direction of the correlation.
}
