% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dominant_topics.R
\name{plot_dominant_topics}
\alias{plot_dominant_topics}
\title{Visualize the dominant topic for each sample in a Seurat object.}
\usage{
plot_dominant_topics(
  object,
  x,
  palette = NULL,
  marker = ".",
  s = 20,
  fontsize = 10,
  markerscale = 2,
  figsize = c(7, 5),
  transparent = FALSE,
  save = NULL
)
}
\arguments{
\item{object}{A Seurat object containing multimodal single-cell data,
including topic probabilities and coordinates.}

\item{x}{A name of the embedding to use for plotting (e.g., UMAP, PCA).
If equal to ’Spatial’ tries to use spot coordinates.}

\item{palette}{A list mapping topics to specific colors. If NULL, a default
palette of unique hex colors is generated. Default is NULL.}

\item{marker}{Marker style for the scatter plot. Default is ‘.‘.}

\item{s}{Marker size in the scatter plot. Default is 20.}

\item{fontsize}{Font size for legend labels. Default is 10.}

\item{markerscale}{Scale of markers in the legend relative to their size in
the scatter plot. Default is 2.}

\item{figsize}{Vector specifying the figure size c(width, height) in inches.
Default is c(7, 5).}

\item{transparent}{Whether to save the figure with a transparent background.
Useful for overlays or presentations. Default is FALSE.}

\item{save}{Path to save the figure. If NULL, the figure is displayed but
not saved. Default is NULL.}
}
\description{
This function creates a scatter plot where each point
 represents a cell/spot, colored according to the dominant topic (i.e., the
 topic with the highest probability) for that sample. The plot provides an
 intuitive overview of how topics are distributed spatially or in a given
 embedding. A legend maps colors to topics for easy interpretation.
}
