% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topics.R
\name{plot_topics}
\alias{plot_topics}
\title{Visualize topic distributions on spatial coordinates or embedding.}
\usage{
plot_topics(
  object,
  x,
  cmap = "gnuplot",
  marker = ".",
  s = 10,
  vmax = NULL,
  fontsize = 10,
  figsize = NULL,
  transparent = FALSE,
  save = NULL
)
}
\arguments{
\item{object}{A Seurat object containing multimodal single-cell data, with
topic distributions stored in ’mTopic’ reduction embedding.}

\item{x}{A name of the embedding to use for plotting (e.g., UMAP, PCA).
If equal to ’Spatial’ tries to use spot coordinates.}

\item{cmap}{Colormap to use for visualizing topic proportions. Default is
‘gnuplot’.}

\item{marker}{Marker style for scatter plots. Default is ‘.’.}

\item{s}{Marker size in scatter plots. Default is 10.}

\item{vmax}{Maximum value for the color scale. If NULL, it is set to the
99.9th percentile of the data for each topic. Default is NULL.}

\item{fontsize}{Font size for plot titles and colorbar ticks. Default is 10.}

\item{figsize}{Vector specifying the figure size c(width, height) in inches.
If NULL, the size is automatically determined based on the number of
topics. Default is NULL.}

\item{transparent}{Whether to make the figure background transparent. Useful
for overlays in presentations. Default is FALSE.}

\item{save}{Path to save the figure. If NULL, the figure is displayed but
not saved. Default is NULL.}
}
\description{
This function generates scatter plots to visualize the
 topic distributions across cells/spots. Each topic is displayed
 in a separate subplot, showing how the topic is spatially distributed or
 distributed within a specified embedding. The plots are arranged in a grid
 for easy comparison.
}
