% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zscores.R
\name{plot_zscores}
\alias{plot_zscores}
\title{Visualize the spatial or embedding-based distribution of z-scores for
 topics in a specified modality.}
\usage{
plot_zscores(
  object,
  mod,
  x,
  cmap = NULL,
  marker = ".",
  s = 10,
  fontsize = 10,
  figsize = NULL,
  transparent = FALSE,
  save = NULL
)
}
\arguments{
\item{object}{A Seurat object containing multimodal single-cell data with
z-scores stored in the misc attribute.}

\item{mod}{The modality to visualize z-scores for (e.g., ‘rna’, ‘protein’).}

\item{x}{The key in the reduction attribute of the Seurat object,
representing the spatial coordinates or embedding to use for plotting
(e.g., ‘umap’, ’Spatial’).}

\item{cmap}{The colormap to use for visualizing z-scores. If NULL, a custom
colormap is applied. Default is NULL.}

\item{marker}{Marker style for scatter plots. Default is ‘.’.}

\item{s}{Marker size for scatter plots. Default is 10.}

\item{fontsize}{Font size for plot titles and colorbar ticks. Default is 10.}

\item{figsize}{Vector specifying the figure size c(width, height) in
inches. If NULL, the size is automatically determined based on the number
of topics. Default is NULL.}

\item{transparent}{Whether to make the figure background transparent. Useful
for overlays in presentations. Default is FALSE.}

\item{save}{Path to save the figure. If NULL, the figure is displayed but
not saved. Default is NULL.}
}
\description{
This function generates scatter plots showing the distribution
 of z-scores for each topic in a specified modality of a Seurat object.
 Z-scores are used to highlight spatial or embedding-based patterns of
 feature expression across samples, revealing significant variations
 relative to the global mean.
}
