% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_counts.R
\name{scale_counts}
\alias{scale_counts}
\title{Scale count matrices to normalize the total sum of counts across
 modalities.}
\usage{
scale_counts(object, counts_per_cell = 10000)
}
\arguments{
\item{object}{A Seurat object containing multiple modalities.}

\item{counts_per_cell}{The target average counts per cell. The total sum of
counts in each modality is scaled to counts_per_cell * n_obs, where n_obs
is the number of cells. Default is 10,000.}
}
\value{
A new Seurat object.
}
\description{
This function normalizes the total sum of counts in the count
 matrix of each modality within the provided Seurat object. Each modality’s
 counts are scaled so that their total sum matches a specified target value
 (counts_per_cell * n_obs). This ensures consistency in total counts across
 different modalities, which is crucial for comparative and integrative
 analyses in single-cell multimodal datasets.
}
\note{
This function is especially useful when working with multimodal
datasets where each modality may have different total counts, making
comparisons challenging.
}
