% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{tfidf}
\alias{tfidf}
\title{Apply Term Frequency-Inverse Document Frequency (TF-IDF)
 transformation to a specific modality in a Seurat object.}
\usage{
tfidf(object, mod)
}
\arguments{
\item{object}{A Seurat object containing multiple modalities, each with
a count matrix to be transformed.}

\item{mod}{The modality to apply the TF-IDF transformation to (e.g.,
‘rna’, ‘protein’).}
}
\value{
A new Seurat object with the TF-IDF-transformed data.
}
\description{
This function performs a TF-IDF transformation on the count
 matrix of the specified modality within the provided Seurat object.
 The TF-IDF transformation adjusts raw counts by considering both term
 frequency (TF) and inverse document frequency (IDF). This enhances
 interpretability by down-weighting common features and emphasizing rare
 ones. The transformation is applied in-place by default or to a copy if
 specified.
}
\note{
The TfidfTransformer from sklearn is used with norm=None and
 smooth_idf=False, meaning the output will not be normalized, and
 inverse document frequency will not be smoothed.

This function assumes the count matrix of the specified modality
 is compatible with TfidfTransformer.

Ideal for preprocessing single-cell data with modalities like RNA
 or ATAC.
}
