% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap}
\alias{umap}
\title{Perform UMAP dimensionality reduction on topic distributions and
 store the results in the MuData object.}
\usage{
umap(
  object,
  umap = "mTopic_umap",
  n_components = 2,
  min_dist = 0.1,
  n_neighbors = 20,
  seed = 2291
)
}
\arguments{
\item{object}{A Seurat object containing the topic proportions in the
scmtm reduction.}

\item{umap}{The key under which the UMAP results will be stored in the
Seurat embedding. Default is mTopic_umap.}

\item{n_components}{The number of dimensions for the UMAP embedding.
Default is 2.}

\item{min_dist}{The minimum distance between points in the UMAP embedding.
Controls the balance between local and global structure. Default is 0.1.}

\item{n_neighbors}{The number of nearest neighbors to consider when
computing the UMAP embedding. Default is 20.}

\item{seed}{Random seed for reproducibility. Ensures consistent embeddings
across runs. Default is 2291.}
}
\value{
A Seurat object.
}
\description{
This function applies Uniform Manifold Approximation and
 Projection (UMAP) to reduce the dimensionality of topic distributions.
}
\note{
UMAP is a non-linear dimensionality reduction technique widely used
 for visualizing high-dimensional data.
}
