% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_assays.R
\name{update_assays}
\alias{update_assays}
\title{Update Assay to Assay5.}
\usage{
update_assays(object)
}
\arguments{
\item{object}{A Seurat object.}
}
\value{
A new Seurat object containing assays of class Assay5.
}
\description{
This function converts every assay of class Assay to Assay5.
}
