% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VI.R
\name{VI}
\alias{VI}
\title{Perform Variational Inference (VI) to infer topics from the data.}
\usage{
VI(model, n_iter = 20, max_iter_d = 100)
}
\arguments{
\item{model}{An MTM or sMTM model.}

\item{n_iter}{Number of iterations for the VI algorithm. Default is 20.}

\item{max_iter_d}{Maximum iterations for the E-step in each VI update.
Controls convergence criteria. Default is 100.}
}
\value{
MTM or sMTM object.
}
\description{
VI is a deterministic approximation method that updates the
 model’s variational parameters over several iterations to optimize its fit
 to the data. Use VI for moderate-sized datasets where the full dataset can
 be used in each iteration.
}
